+++ web/usuaris/src/helpers/index.ts
@@ -4,3 +4,5 @@ export {
   loginUserFetcher,
   patchUser,
 } from './fetchers'
+
+export { buildQueryString } from './filters'